+++ nym-vpn-app/src/types/tauri.ts
@@ -349,6 +349,7 @@ export type TAccountSummary = {
   trafficUsedGb: bigint;
   trafficLimitGb: bigint;
   trafficResetTime: bigint | null;
+  fairUsageDataUnavailable: boolean;
   accountAddr: string;
   canonicalAccountAddr: string | null;
   authMethods: Array<TAuthMethod>;