+++ nym-vpn-app/CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow adding custom apps (user selected) to split tunneling for Windows & Linux
 - Mixnet tuning settings
 
+### CHANGED
+
+- Switch to daily allowance usage from monthly
+
 ## [1.30.0] - 2026-05-29
 
 ### Fixed