+++ nym-vpn-app/CHANGELOG.md
@@ -7,17 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [2026.10.0] - 2026-06-09
-
 ### Added
 
-- Allow adding custom apps (user selected) to split tunneling for Windows & Linux
 - Add new "Geo exclusion" settings to allow bypass the VPN tunnel when accessing from certain geographic regions. Currently only China is supported
 - Mixnet tuning settings
 
 ### CHANGED
 
 - Switch to daily allowance usage from monthly
+
+## [2026.10.0] - 2026-06-09
+
+### Added
+
+- Allow adding custom apps (user selected) to split tunneling for Windows & Linux
 
 ## [1.30.0] - 2026-05-29